PRO-4812: Moved validation into core.
