+++ build.gradle
@@ -38,7 +38,7 @@ mainClassName = 'uk.gov.hmcts.probate.SubmitApplication'
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
 
-check.dependsOn dependencyCheckAnalyze
+//check.dependsOn dependencyCheckAnalyze
 
 jacoco {
   toolVersion = '0.8.0'
@@ -188,7 +188,8 @@ def versions = [
         springfoxSwagger                : '2.9.2',
         springCloudWiremock             : '1.2.4.RELEASE',
         serentity                       : '2.0.17',
-        pact_version                    : '3.5.24'
+        pact_version                    : '3.5.24',
+        probateCommons                  : '0.0.36-dev'
 ]
 
 
@@ -220,7 +221,7 @@ dependencies {
   compile group: 'uk.gov.hmcts.reform', name: 'core-case-data-store-client', version: versions.ccdClient
   compile group: 'uk.gov.hmcts.reform', name: 'health-spring-boot-starter', version: '0.0.3'
   compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.8.1'
-  compile group: 'uk.gov.hmcts.reform.probate', name: 'probate-commons', version: '0.0.36-dev'
+  compile group: 'uk.gov.hmcts.reform.probate', name: 'probate-commons', version: versions.probateCommons
 
   compile group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0.1'
   compile group: 'org.projectlombok', name: 'lombok', version: '1.16.18'
