+++ src/main/java/uk/gov/hmcts/probate/services/submit/core/validation/Validation.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.services.submit.validation;
+package uk.gov.hmcts.probate.services.submit.core.validation;
 
 @FunctionalInterface
 public interface Validation<K> {
