+++ src/main/java/uk/gov/hmcts/probate/services/submit/core/validation/ValidationResult.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.services.submit.validation;
+package uk.gov.hmcts.probate.services.submit.core.validation;
 
 
 public class ValidationResult {
