+++ src/main/java/uk/gov/hmcts/probate/services/submit/core/validation/ValidationRule.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.services.submit.validation;
+package uk.gov.hmcts.probate.services.submit.core.validation;
 
 import uk.gov.hmcts.reform.probate.model.cases.CaseData;
