+++ src/main/java/uk/gov/hmcts/probate/services/submit/core/validation/rules/CaveatRules.java
@@ -1,9 +1,9 @@
-package uk.gov.hmcts.probate.services.submit.validation.rules;
+package uk.gov.hmcts.probate.services.submit.core.validation.rules;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import uk.gov.hmcts.probate.services.submit.validation.ValidationRule;
+import uk.gov.hmcts.probate.services.submit.core.validation.ValidationRule;
 import uk.gov.hmcts.reform.probate.model.cases.caveat.CaveatData;
 
 @Configuration
@@ -16,5 +16,4 @@ public class CaveatRules {
                         caveat.getDeceasedDateOfDeath().isBefore(caveat.getDeceasedDateOfBirth())
                 , "DeceasedDateOfDeath before DeceasedDateOfBirth");
     }
-
 }
