+++ src/main/java/uk/gov/hmcts/probate/services/submit/core/validation/rules/PaRules.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.services.submit.validation.rules;
+package uk.gov.hmcts.probate.services.submit.core.validation.rules;
 
 import org.springframework.context.annotation.Configuration;
