+++ src/main/java/uk/gov/hmcts/probate/services/submit/core/validation/rules/ValidatorUtils.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.services.submit.validation.rules;
+package uk.gov.hmcts.probate.services.submit.core.validation.rules;
 
 public final class ValidatorUtils {
