+++ src/test/java/uk/gov/hmcts/probate/services/submit/core/validation/GrantOfRepresentationCreator.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.services.submit.validation;
+package uk.gov.hmcts.probate.services.submit.core.validation;
 
 import com.google.common.collect.Lists;
 import uk.gov.hmcts.reform.probate.model.IhtFormType;
